// api/src/helpers/app-error-code.enum.ts
export enum AppErrorCode {
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  FILE_REQUIRED = 'FILE_REQUIRED',
  INVALID_FILE_EXTENSION = 'INVALID_FILE_EXTENSION',
  INVALID_FILE_MIMETYPE = 'INVALID_FILE_MIMETYPE',
  FILE_TOO_LARGE = 'FILE_TOO_LARGE',
  UNEXPECTED_FILE_FIELD = 'UNEXPECTED_FILE_FIELD',
  MALFORMED_MULTIPART = 'MALFORMED_MULTIPART',

  API_KEY_MISSING = 'API_KEY_MISSING',
  API_KEY_INVALID = 'API_KEY_INVALID',

  SESSION_ALREADY_EXISTS = 'SESSION_ALREADY_EXISTS',
  GCS_PERMISSION_DENIED = 'GCS_PERMISSION_DENIED',
  GCS_BUCKET_NOT_FOUND = 'GCS_BUCKET_NOT_FOUND',
  GCS_UPLOAD_FAILED = 'GCS_UPLOAD_FAILED',

  INTERNAL_ERROR = 'INTERNAL_ERROR',
}